Explictly throw an exception if user message class is not public.
https://github.com/blehnen/DotNetWorkQueue/issues/18

The underlying issue is that dynamic is used to avoid having to pass <TMessage> everywhere in the consumer;  However, that has a drawback;  The queue is unable to use dyanmic and invoke the user action delegate if the message class is internal/private.

This may be changed in the future, but the change is not quite straight forward;  We would rather avoid forcing the ConsumerQueue* to be Typed to the message, but it holds a reference to a message processor that would chain all of the way down; This would need to be thought out a bit better.